Refactor form layout
Redid the form so that when a row is clicked, an accordion slides down
under the row with an edit button.  when the edit button is clicked,
the current row is replaced with an edit form.  also, a confirm edit
button shows which will actually make the edit, then make an index
call again.

Next: get update actually working