Update functionality working
Initially my update form had a table cell for each input, which wasn't
allowing the form to submit correctly.  Fixed by wrapping the entire
form in a table cell with column length of 4.  Made the update form
hidden on get index, and toggles with the current row that is clicked.